Refactor project architecture and add Makefile
